[tests] remove bignum.py
It only contains one function and is only imported by one
other module (script.py). Just move the function to
script.py.
